gui/Python/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
FreeSimpleGUI==5.1.1
greenlet==3.1.1
idna==3.10
keyboard==0.13.5
Nuitka==2.5.1
ordered-set==4.1.0
packaging==24.2
pefile==2023.2.7
proxmoxer==2.2.0
pyinstaller-hooks-contrib==2025.1
PySide6==6.8.0
PySide6_Addons==6.8.0
PySide6_Essentials==6.8.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
shiboken6==6.8.0
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
